Added navi and changed background
