+++ src/App.css
@@ -19,21 +19,21 @@
     animation: App-logo-spin infinite 20s linear;
   }
 }
-@media (prefers-reduced-motion: no-preference) {
-.App-header {
-  background-image: url("acc.jpg");
-  background-size: cover;
-  height: 400px;
-  width: 100%;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(60px + 2vmin);
-  color: white;
-}
-}
+/*@media (prefers-reduced-motion: no-preference) {*/
+/*.App-header {*/
+/*  background-image: url("acc.jpg");*/
+/*  background-size: cover;*/
+/*  height: 400px;*/
+/*  width: 100%;*/
+/*  min-height: 100vh;*/
+/*  display: flex;*/
+/*  flex-direction: column;*/
+/*  align-items: center;*/
+/*  justify-content: center;*/
+/*  font-size: calc(60px + 2vmin);*/
+/*  color: white;*/
+/*}*/
+/*}*/
 .App-link {
   color: #61dafb;
   padding: 10px 20px;
