+++ src/navigation/styles.css
@@ -1,0 +1,52 @@
+.App {
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+  background-color: #282c34;
+  font-size: calc(10px + 2vmin);
+  color: white;
+  min-height: 50vh;
+  align-items: center;
+}
+
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
+@media (prefers-reduced-motion: no-preference) {
+.Navigation {
+  background-color: black;
+  background-image: url("https://acegif.com/wp-content/gif/outerspace-30.gif");
+  background-size: 60% 100%;
+  background-repeat: no-repeat;
+  background-position: center;
+  height: 400px;
+  width: 100%;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: start;
+  justify-content: center;
+  font-size: calc(20px + 2vmin);
+  color: white;
+}
+}
+/*.App-link {*/
+/*  color: #61dafb;*/
+/*  padding: 10px 20px;*/
+/*}*/
+
+/*@keyframes App-logo-spin {*/
+/*  from {*/
+/*    transform: rotate(0deg);*/
+/*  }*/
+/*  to {*/
+/*    transform: rotate(360deg);*/
+/*  }*/
+/*}*/
