+++ src/routes/index.ts
@@ -1,0 +1,7 @@
+export const routesPaths = {
+    home: '/',
+    about: 'about',
+    form: 'form',
+    policy: 'policy'
+
+}
